requirement.txt
asgiref==3.2.10
astroid==2.3.3
attrs==19.3.0
automium==0.2.6
automium-web==0.1.1
backcall==0.2.0
bleach==3.1.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.44.0
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-froala-editor==3.2.2
django-js-asset==1.2.2
djangorestframework==3.12.4
entrypoints==0.3
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
ipykernel==5.3.4
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
legacy==0.1.6
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.4
mysql-connector-python==8.0.23
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.1
oauthlib==3.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==1.0.18
pycparser==2.20
Pygments==2.6.1
PyJWT==2.0.1
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2017.3
pywin32==228
pywinpty==0.5.7
PyYAML==3.12
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.23.1
scipy==1.5.2
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.13.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.26.3
views==0.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.11.2
